Create `webhook_deploy` with a path-based `_id`
Hey Nick 👋 

First of all, congratulations for this beautiful, intuitive plugin. Effortless and extremely practical, it's really helping :D 

I'm a bit concerned about security, though: as we need to add a Vercel token & webhook to our configuration, it'd be ideal if the `webhook_deploy` documents created were private. To do this, all we need to do is add a path before the `_id` of each created document, as path-based `_id`s are private by default ([see docs](https://www.sanity.io/docs/ids#paths-fdc25ada5db2)). This makes no difference in private datasets, though.

This little PR adds a cheap and dirty way for enabling this. I opted for Math.random() as it's the cheapest way to add almost-unique ids, but we could always reach for [nanoid](https://www.npmjs.com/package/nanoid) or similar 😉 

Again, thank you a ton  - no rush to review this 😄