Update README.md
The Team Name field refers to the editable team name of the user's Vercel Team. In order to get it to work, the slugified Team URL is required - not the Team Name. This is available under Settings > General > Team URL. 

Providing the Team Name returned an error, while providing the Team URL did the trick